chore(ui): move deprecated components, add children type in story, improve types
