+++ packages/ui-components/src/components/FormattedText/FormattedText.stories.tsx
@@ -14,6 +14,9 @@ export default {
   argTypes: {
     children: {
       control: false,
+      table: {
+        type: { summary: "ReactNode" },
+      },
     },
   },
   parameters: {
