+++ packages/ui-components/src/components/PageFooter/PageFooter.stories.tsx
@@ -13,6 +13,9 @@ export default {
   argTypes: {
     children: {
       control: false,
+      table: {
+        type: { summary: "ReactNode" },
+      },
     },
   },
 }
