+++ packages/ui-components/src/components/Stack/Stack.stories.tsx
@@ -14,6 +14,9 @@ export default {
   argTypes: {
     children: {
       control: false,
+      table: {
+        type: { summary: "ReactNode" },
+      },
     },
     gap: {
       options: [
