+++ packages/ui-components/src/components/Tab/Tab.stories.tsx
@@ -16,6 +16,9 @@ export default {
     },
     children: {
       control: false,
+      table: {
+        type: { summary: "ReactNode" },
+      },
     },
   },
 }
