+++ packages/ui-components/src/components/TabPanel/TabPanel.stories.tsx
@@ -11,6 +11,9 @@ export default {
   argTypes: {
     children: {
       control: false,
+      table: {
+        type: { summary: "ReactNode" },
+      },
     },
   },
 }
