+++ packages/ui-components/src/index.ts
@@ -26,7 +26,6 @@ export { ContentHeading } from "./components/ContentHeading/ContentHeading.compo
 export { ContentAreaToolbar } from "./components/ContentAreaToolbar/ContentAreaToolbar.component"
 export { ContentContainer } from "./components/ContentContainer/ContentContainer.component"
 export { Container } from "./components/Container/Container.component"
-export { ContextMenu } from "./components/ContextMenu/ContextMenu.component"
 export { DataGrid } from "./components/DataGrid/DataGrid.component"
 export { DataGridCell } from "./components/DataGridCell/DataGridCell.component"
 export { DataGridHeadCell } from "./components/DataGridHeadCell/DataGridHeadCell.component"
@@ -50,9 +49,6 @@ export { LoadingIndicator } from "./components/LoadingIndicator/LoadingIndicator
 export { MainContainer } from "./components/MainContainer/MainContainer.component"
 export { MainContainerInner } from "./components/MainContainerInner/MainContainerInner.component"
 export { MainTabs } from "./components/MainTabs//MainTabs.component"
-export { Menu } from "./components/Menu/Menu.component"
-export { MenuItem } from "./components/MenuItem/MenuItem.component"
-export { MenuSection } from "./components/MenuSection/MenuSection.component"
 export { Message } from "./components/Message/Message.component"
 export { Modal } from "./components/Modal/Modal.component"
 export { ModalFooter } from "./components/ModalFooter/ModalFooter.component"
@@ -102,6 +98,7 @@ export { useEndlessScrollList } from "./hooks/useEndlessScrollList"
 export { AppIntro } from "./deprecated_js/AppIntro/AppIntro.component"
 export { CheckboxRow } from "./deprecated_js/CheckboxRow"
 export { ContentArea } from "./deprecated_js/ContentArea/ContentArea.component"
+export { ContextMenu } from "./deprecated_js/ContextMenu/ContextMenu.component"
 export { ContentAreaWrapper } from "./deprecated_js/ContentAreaWrapper/ContentAreaWrapper.component"
 export { DataList } from "./deprecated_js/DataList/DataList.component"
 export { DataListCell } from "./deprecated_js/DataListCell/DataListCell.component"
@@ -110,6 +107,9 @@ export { DataListRow } from "./deprecated_js/DataListRow/DataListRow.component"
 export { FilterInput } from "./deprecated_js/FilterInput/FilterInput.component"
 export { FilterPill } from "./deprecated_js/FilterPill/FilterPill.component"
 export { Filters } from "./deprecated_js/Filters/Filters.component"
+export { Menu } from "./deprecated_js/Menu/Menu.component"
+export { MenuItem } from "./deprecated_js/MenuItem/MenuItem.component"
+export { MenuSection } from "./deprecated_js/MenuSection/MenuSection.component"
 export { RadioRow } from "./deprecated_js/RadioRow"
 export { SelectRow } from "./deprecated_js/SelectRow"
 export { SwitchRow } from "./deprecated_js/SwitchRow"
